Added link to responsive page from Full pages
Added link in footer that goes back to default (responsive) page so that
if someone goes to the "Full" site, they can get back to the responsive
site if they want. Changed all links on "Full" pages to link to other
"Full" pages, so users won't have to navigate to the "Full" page more
than once.